requirements.txt
pytest
coverage
RPi.GPIO
pygame

